backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
# [MIGRATION] 切换到 Google 官方新版 SDK
google-genai==0.3.0
python-dotenv==1.0.1
httpx==0.26.0